Enforce float NaN imputation contract in ML
Standardized training and scoring inputs to use float-typed DataFrames so missing values become NaN and are handled consistently by the pipeline imputers. `_rows_to_matrix` and `_extract_drivers` now both construct `dtype=float` frames, aligning production scoring with training-time behavior. Added a regression test that loads the saved pipeline and verifies advisory-family features still impute to 0 while non-advisory features (e.g., GitHub stars) retain median imputation.

+++ canary/scoring/ml.py
@@ -384,7 +384,9 @@ def _extract_drivers(
 
     # Impute values the same way the pipeline does so contributions are accurate
     try:
-        X_raw = pd.DataFrame([raw_values], columns=feature_columns)  # pyright: ignore[reportArgumentType]
+        X_raw = pd.DataFrame(  # pyright: ignore[reportArgumentType]
+            [raw_values], columns=feature_columns, dtype=float
+        )  # float64 contract: None -> NaN so the imputers mask it (see baseline.py)
         X_imputed: Any = imputer_step.transform(X_raw) if imputer_step is not None else X_raw.values
     except Exception as exc:
         LOGGER.debug("Imputation failed in driver extraction; using zero-fill.", exc_info=exc)
